Refactor JavaEmitter
