+++ src/compiler/emitters/java/src/commonMain/kotlin/community/flock/wirespec/emitters/java/JavaEnumDefinitionEmitter.kt
@@ -1,0 +1,33 @@
+package community.flock.wirespec.emitters.java
+
+import community.flock.wirespec.compiler.core.emit.EnumDefinitionEmitter
+import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
+import community.flock.wirespec.compiler.core.emit.SpaceEmitter
+import community.flock.wirespec.compiler.core.emit.Spacer
+import community.flock.wirespec.compiler.core.parse.Enum
+import community.flock.wirespec.compiler.core.parse.Module
+
+interface JavaEnumDefinitionEmitter: JavaIdentifierEmitter, EnumDefinitionEmitter, SpaceEmitter {
+
+    override fun emit(enum: Enum, module: Module) = """
+        |public enum ${emit(enum.identifier)} implements Wirespec.Enum {
+        |${enum.entries.joinToString(",\n") { "${it.sanitizeEnum().sanitizeKeywords()}(\"$it\")" }.spacer()};
+        |${Spacer}public final String label;
+        |${Spacer}${emit(enum.identifier)}(String label) {
+        |${Spacer(2)}this.label = label;
+        |${Spacer}}
+        |${Spacer}@Override
+        |${Spacer}public String toString() {
+        |${Spacer(2)}return label;
+        |${Spacer}}
+        |${Spacer}@Override
+        |${Spacer}public String getLabel() {
+        |${Spacer(2)}return label;
+        |${Spacer}}
+        |}
+        |
+    """.trimMargin()
+
+    fun String.sanitizeEnum() = split("-", ", ", ".", " ", "//").joinToString("_").sanitizeFirstIsDigit()
+
+}
