+++ src/compiler/emitters/java/src/commonMain/kotlin/community/flock/wirespec/emitters/java/JavaIdentifierEmitter.kt
@@ -1,0 +1,32 @@
+package community.flock.wirespec.emitters.java
+
+import community.flock.wirespec.compiler.core.addBackticks
+import community.flock.wirespec.compiler.core.emit.Emitter.Companion.firstToUpper
+import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
+import community.flock.wirespec.compiler.core.parse.DefinitionIdentifier
+import community.flock.wirespec.compiler.core.parse.FieldIdentifier
+import community.flock.wirespec.compiler.core.parse.Identifier
+import community.flock.wirespec.emitters.java.JavaEmitter.Companion.reservedKeywords
+import community.flock.wirespec.emitters.kotlin.KotlinEmitter.Companion.reservedKeywords
+
+interface JavaIdentifierEmitter: IdentifierEmitter {
+
+    override fun emit(identifier: Identifier) = when (identifier) {
+        is DefinitionIdentifier -> identifier.value.sanitizeSymbol()
+        is FieldIdentifier -> identifier.value.sanitizeSymbol().sanitizeKeywords()
+    }
+
+    fun String.sanitizeSymbol() = this
+        .split(".", " ", "-")
+        .mapIndexed { index, s -> if (index > 0) s.firstToUpper() else s }
+        .joinToString("")
+        .asSequence()
+        .filter { it.isLetterOrDigit() || it in listOf('_') }
+        .joinToString("")
+        .sanitizeFirstIsDigit()
+
+    fun String.sanitizeFirstIsDigit() = if (firstOrNull()?.isDigit() == true) "_${this}" else this
+
+    fun String.sanitizeKeywords() = if (this in reservedKeywords) "_$this" else this
+
+}
